Version 3, initial commit
Version 2 is highly verbose in command line options. This new implementation aims to reduce the command lengths by using config files.

It also assumes the hosts are managed by the CA, so keeps track of keys and certs locally.

Generic CSR gen and CSR sign are also to be supported.